separated controller and application for plant form.
