+++ src/main/java/com/example/alchemillafx/PlantController.java
@@ -1,20 +1,24 @@
 package com.example.alchemillafx;
 
-import javafx.application.Application;
+// import javafx.application.Application;
 import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
-import javafx.scene.layout.VBox;
-import javafx.stage.Stage;
+// import javafx.fxml.FXMLLoader;
+// import javafx.fxml.Initializable;
+// import javafx.scene.Scene;
+// import javafx.scene.control.Label;
+// import javafx.scene.layout.VBox;
+// import javafx.stage.Stage;
 import javafx.scene.control.TextField;
 import javafx.scene.text.*;
 
-import java.awt.image.BufferedImage;
+// import java.awt.image.BufferedImage;
 import java.io.*;
-import javax.imageio.ImageIO;
+// import javax.imageio.ImageIO;
+import javafx.event.ActionEvent;
 
-public class PlantController extends Application
+import javax.swing.*;
+
+public class PlantController
 {
     @FXML
     private Text actiontarget;
@@ -44,26 +48,10 @@ public class PlantController extends Application
     Profile profile;
     String currentProfile = "joe";
 
-    @Override
-    public void start(Stage primaryStage) throws Exception
-    {
-        // Sets Title, Loads the FXML File
-        primaryStage.setTitle("Enter Plant Info");
-        FXMLLoader loader = new FXMLLoader();
-        String path = "@prj-03-Alchemilla/src/resources/com.example.alchemilla/plant-form.fxml";
-        FileInputStream stream = new FileInputStream(path);
-
-        // we vboxed up; creates pane and details
-        VBox box = new VBox(loader.load(stream));
-
-        Scene scene = new Scene(box, 2000, 2000); // arbitrary numbers
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
-
     @FXML
-    protected void onButtonClick() {
-        actiontarget.setText("Welcome to JavaFX Application!");
+    protected void onButtonClick(ActionEvent ae) {
+        // probably want to have something here that makes sure that all data entries are there from UI
+        //  actiontarget.setText("Welcome to JavaFX Application!");
 
         String imageString = plantimagepath.getText();
         String nameString = plantname.getText();
@@ -76,6 +64,7 @@ public class PlantController extends Application
 
         Plant plant = new Plant(imageString, nameString, Integer.parseInt(waterString), Integer.parseInt(foodDeltaString), foodString, descString, instructionString);
 
+        // Save to profile
         SerializeClassData sc = new SerializeClassData();
         try {
             profile = sc.getProfile(currentProfile);
@@ -107,8 +96,4 @@ public class PlantController extends Application
         return image;
     }
 */
-    public static void main (String[] args)
-    {
-        Application.launch(args);
-    }
 }
